software/src/requirements.txt
pandas==2.2.3
scanpy==1.10.1
numpy==2.2.6
scipy==1.15.3
scikit-learn==1.6.1
harmonypy==0.0.9
polars-lts-cpu==1.32.1